analytics/requirements.txt
kafka-python==2.0.2
PyMySQL==1.0.2
user-agents==2.2.0